requirements.txt
absl-py==0.12.0
autopep8==1.5.6
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.3
six==1.15.0
sphfile==1.0.3
toml==0.10.2
torch==1.8.1
torchaudio==0.8.1
typing-extensions==3.7.4.3
